Added a form to display the results
Added a form to display the results of the method ans what weight plates
are to be placed on the bar.

Updated the weightcalculator method to include a display of whatever
weight is left over if a number that isnt cleanly divided by the plates
is entered, so that the user knows if the full weight they entered isnt
fully there.